Improve accessibility and password field behavior
Marked fields that are required with an asterisk and added labels
to each asterisk so screen readers can recognize them.

The form now validates password fields on submit.

Added two images below each of the password field's labels for showing
and hiding the fields.

Added a message below each of the password fields that let user know
how to validate them if necessary.